Added all of the webpage code
This push includes the html, css, new routes, and some minor bug fixes
that were needed to get everything up and running. In this state the
system works with the 'simple client' and it is ready to be deployed.

TODO: add the ability to request images and display instead of just
receiving images, although this might require architecture changes and
separate flask server apps on the cameras to handle GET requests.